### 2022-04-13 #171 @whedon
[PRE REVIEW]: What and How of Machine Learning Transparency: Building Bespoke Explainability Tools with Interoperable Algorithmic Components
**Submitting author:** @So-Cool (<a href="http://orcid.org/0000-0002-9869-5896">Kacper Sokol</a>)
**Repository:** <a href="https://github.com/fat-forensics/Surrogates-Tutorial" target ="_blank">https://github.com/fat-forensics/Surrogates-Tutorial</a>
**Version:** 2020-ecml-pkdd
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
