### 2022-04-13 #171 @labarba
Great! Here's a neat trick: you can run `@whedon assign me as editor`
