### 2022-04-22 ef6ddc1
BibTeX DOIs and JOSE badge
