### 2022-05-03 #171 @ttimbers
@So-Cool - thank-you for your submission. I have read your manuscript and considered it with the scope of JOSE. Typically, short learning modules are not within scope of the journal. However, short learning modules can be considered if the authors show substantial scholarly effort in other ways. One example of this would be if the tutorial has been given many times and iteratively improved. Another example would be that the module has been subject to methodical evaluation. These scholarly efforts should be reported in the manuscript.

Reading your manuscript, I did note you mentioned this module has been taught more than once, and perhaps by others? If this were further elaborated on, as well as what has been learned from teaching this several times, that could be considered on of the scholarly efforts we described above.
