+++ paper/paper.bib
@@ -117,3 +117,15 @@
   year={2001},
   publisher={JSTOR}
 }
+
+@inproceedings{sokol2021you,
+  title={You Only Write Thrice:
+         {C}reating Documents, Computational Notebooks and Presentations
+         From a Single Source},
+  author={Sokol, Kacper and Flach, Peter},
+  booktitle={Beyond static papers: {R}ethinking how we share scientific
+             understanding in Machine Learning},
+  series={ICLR 2021 Workshop},
+  doi={10.48550/arXiv.2107.06639},
+  year={2021}
+}
