+++ paper/paper.bib
@@ -125,7 +125,7 @@
   author={Sokol, Kacper and Flach, Peter},
   booktitle={Beyond static papers: {R}ethinking how we share scientific
              understanding in Machine Learning},
-  series={ICLR 2021 Workshop},
+  publisher={ICLR 2021 Workshop},
   doi={10.48550/arXiv.2107.06639},
   year={2021}
 }
