+++ paper/paper.bib
@@ -124,8 +124,20 @@
          From a Single Source},
   author={Sokol, Kacper and Flach, Peter},
   booktitle={Beyond static papers: {R}ethinking how we share scientific
-             understanding in Machine Learning},
-  publisher={ICLR 2021 Workshop},
+             understanding in Machine Learning -- {ICLR} 2021 Workshop},
   doi={10.48550/arXiv.2107.06639},
   year={2021}
+}
+
+@article{rudin2019stop,
+  title={Stop explaining black box machine learning models for high stakes
+         decisions and use interpretable models instead},
+  author={Rudin, Cynthia},
+  journal={Nature Machine Intelligence},
+  doi={10.1038/s42256-019-0048-x},
+  volume={1},
+  number={5},
+  pages={206--215},
+  year={2019},
+  publisher={Nature Publishing Group}
 }
