+++ paper/paper.bib
@@ -80,3 +80,64 @@
   doi={10.48550/arXiv.2112.14466},
   year={2021}
 }
+
+@article{breiman2001random,
+  title={Random Forests},
+  author={Breiman, Leo},
+  journal={Machine Learning},
+  doi={10.1023/A:1010933404324},
+  volume={45},
+  number={1},
+  pages={5--32},
+  year={2001},
+  publisher={Springer}
+}
+
+@article{goldstein2015peeking,
+  title={Peeking inside the black box:
+         {V}isualizing statistical learning with plots of individual
+         conditional expectation},
+  author={Goldstein, Alex and Kapelner, Adam and Bleich, Justin and
+          Pitkin, Emil},
+  journal={Journal of Computational and Graphical Statistics},
+  doi={10.1080/10618600.2014.907095},
+  volume={24},
+  number={1},
+  pages={44--65},
+  year={2015},
+  publisher={Taylor \& Francis}
+}
+
+@article{friedman2001greedy,
+  title={Greedy function approximation: {A} gradient boosting machine},
+  author={Friedman, Jerome H},
+  journal={Annals of statistics},
+  doi={10.1214/aos/1013203451},
+  pages={1189--1232},
+  year={2001},
+  publisher={JSTOR}
+}
+
+@inproceedings{sokol2021you,
+  title={You Only Write Thrice:
+         {C}reating Documents, Computational Notebooks and Presentations
+         From a Single Source},
+  author={Sokol, Kacper and Flach, Peter},
+  booktitle={Beyond static papers: {R}ethinking how we share scientific
+             understanding in Machine Learning -- {ICLR} 2021 Workshop},
+  doi={10.48550/arXiv.2107.06639},
+  year={2021}
+}
+
+@article{rudin2019stop,
+  title={Stop explaining black box machine learning models for high stakes
+         decisions and use interpretable models instead},
+  author={Rudin, Cynthia},
+  journal={Nature Machine Intelligence},
+  doi={10.1038/s42256-019-0048-x},
+  volume={1},
+  number={5},
+  pages={206--215},
+  year={2019},
+  publisher={Nature Publishing Group}
+}
