### 2022-05-11 #171 @So-Cool
Thank you for your review, @ttimbers. We have revised the manuscript based on your comments. In particular, the last paragraph of the "Modular Surrogate Explainers" section has been updated and extended by a new paragraph explaining the evolution of the training materials and our teaching experience.
