### 2022-05-30 #171 @ttimbers
@kbodwin - hi! 👋 you are listed on the JOSE reviewer volunteer list, and that you have domain expertise in machine learning. Thus, I wonder if you would consider contributing a review on this manuscript submitted to JOSE. We are in need of some help with this submission:

Title: What and How of Machine Learning Transparency:
Building Bespoke Explainability Tools with Interoperable
Algorithmic Components
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00171/jose.00171/10.21105.jose.00171.pdf)

Thanks for considering it!
