### 2022-06-20 #171 @ttimbers
@mnarayan - thank-you for the offer, however we had another reviewer provide the same offer a bit earlier than you. We will of course keep you in mind for the next machine learning paper we need reviewed at JOSE!
