### 2022-06-20 #175 @whedon
[REVIEW]: What and How of Machine Learning Transparency: Building Bespoke Explainability Tools with Interoperable Algorithmic Components
**Submitting author:** @So-Cool (<a href="http://orcid.org/0000-0002-9869-5896">Kacper Sokol</a>)
**Repository:** <a href="https://github.com/fat-forensics/Surrogates-Tutorial" target ="_blank">https://github.com/fat-forensics/Surrogates-Tutorial</a>
**Version:** 2020-ecml-pkdd

- [ ] **Version:** Does the release version given match the repository release (2020-ecml-pkdd)?

- [ ] **Version:** Does the release version given match the repository release (2020-ecml-pkdd)?
