### 2022-07-06 532f761
Contribution guidelines (fixes #2)
