### 2022-07-06 #175 @So-Cool
@arokem, the new version of the paper addresses both of your comments. [CONTRIBUTING.md](https://github.com/fat-forensics/Surrogates-Tutorial/blob/master/CONTRIBUTING.md) was added to the repository, and the *Statement of Need* section of the paper was expanded to include the missing information.
