### 2022-08-01 #175 @kvarada
@So-Cool Thank you for putting together this excellent resource on this important topic. I enjoyed going through this material.

Here are some of my initial comments:
- The content is well motivated, well structured and easy to follow. It includes many relatable and effective examples. I like that the material has a good balance of theory and application.
- I like that you put emphasize on the fact that the available tools need to be tuned for the given context and scenario. It might be useful to clearly show what can go wrong with one-size-fits all approach. For example, what can go wrong if I just go with SHAP explanations without carefully considering, say, interpretable representations.
- I think the material would benefit from including more non-toy examples.
- In general, it would be helpful for the reader if the README has more information, for example, on the license or installation.
- Do you have the same license for the code and the teaching material?
- I couldn't find paper.md file in the GitHub repository.

And here I'm copying the checklists, as I'm not able to tick them off above.

- [x] Repository: Is the source for this learning module available at the [repository url](https://github.com/fat-forensics/Surrogates-Tutorial)?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)
- [x] Version: Does the release version given match the repository release (2020-ecml-pkdd)?
- [x] Authorship: Has the submitting author (@So-Cool) made visible contributions to the module? Does the full list of authors seem appropriate and complete?

- [x] A statement of need: Do the authors clearly state the need for this module and who the target audience is?
- [x] Installation instructions: Is there a clearly stated list of dependencies?
- [x] Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?

- [x] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

- [x] Learning objectives: Does the module make the learning objectives plainly clear? (We don't require explicitly written learning objectives; only that they be evident from content and design.)
- [x] Content scope and length: Is the content substantial for learning a given topic? Is the length of the module appropriate?
- [x] Pedagogy: Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)
- [x] Content quality: Is the writing of good quality, concise, engaging? Are the code components well crafted? Does the module seem complete?
- [x] Instructional design: Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.

- [ ] Authors: Does the paper.md file include a list of authors with their affiliations?
- [x] A statement of need: Does the paper clearly state the need for this module and who the target audience is?
- [x] Description: Does the paper describe the learning materials and sequence?
- [x] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
