### 2022-08-05 d5850de
Improving README, licencing information and references
