+++ README.md
@@ -1,30 +1,72 @@
 [![Read More](https://img.shields.io/badge/read-more-blue)](https://events.fat-forensics.org/2020_ecml-pkdd)
-[![Tests Status](https://github.com/fat-forensics/Surrogates-Tutorial/actions/workflows/tests.yml/badge.svg)](https://github.com/fat-forensics/Surrogates-Tutorial/actions/workflows/tests.yml)
-[![Licence](https://img.shields.io/badge/licence-CC%20BY--NC--SA%204.0-red)](LICENCE)  
-[![ZENODO DOI](https://zenodo.org/badge/472186940.svg)](https://zenodo.org/badge/latestdoi/472186940)
+[![Text Licence](https://img.shields.io/badge/licence--text-CC%20BY--NC--SA%204.0-red)](LICENCE)
+[![Code Licence](https://img.shields.io/badge/licence--code-new%20BSD-red)](LICENCE-code)  
 [![JOSE DOI](https://jose.theoj.org/papers/d58625bd4c600da866522c879986b18f/status.svg)](https://jose.theoj.org/papers/d58625bd4c600da866522c879986b18f)
+[![ZENODO DOI](https://zenodo.org/badge/472186940.svg)](https://zenodo.org/badge/latestdoi/472186940)
 [![GitHub Release](https://img.shields.io/github/v/release/fat-forensics/Surrogates-Tutorial?display_name=tag&logo=github)](https://github.com/fat-forensics/Surrogates-Tutorial/releases/latest)  
+[![Tests Status](https://github.com/fat-forensics/Surrogates-Tutorial/actions/workflows/tests.yml/badge.svg)](https://github.com/fat-forensics/Surrogates-Tutorial/actions/workflows/tests.yml)  
 [![Open in Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/fat-forensics/Surrogates-Tutorial/master?filepath=notebooks)
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/fat-forensics/Surrogates-Tutorial/blob/master/)
 
 # What and How of Machine Learning Transparency #
 ## Building Bespoke Explainability Tools with Interoperable Algorithmic Components ##
 
-> *An ECML-PKDD 2020 Hands-on Tutorial.*
+Explainability techniques for data-driven predictive models based on
+artificial intelligence and machine learning algorithms allow us to
+better understand the operation of such systems and hold them
+accountable[^1].
+New transparency approaches are therefore developed at breakneck speed
+to peek inside these black boxes and interpret their decisions.
+Many of these techniques are introduced as monolithic tools, giving the
+impression of one-size-fits-all and end-to-end algorithms with limited
+customisability.
+However, such approaches are often composed of multiple interchangeable
+modules that need to be tuned to the problem at hand to produce meaningful
+explanations[^2].
+This repository holds a collection of interactive, hands-on training materials
+(offered as Jupyter Notebooks) that provide guidance through the process of
+building and evaluating bespoke modular surrogate explainers for black-box
+predictions of tabular data.
+These resources cover the three core building blocks of this technique
+introduced by the bLIMEy meta-algorithm: interpretable representation
+composition, data sampling and explanation generation[^2].
 
-This repository holds a collection of interactive training materials on
-building bespoke surrogate explainers of individual black-box predictions
-for tabular data.
-The following resources are available:
+The following materials are available
+(follow the links to see their respective descriptions):
 
[…]
