### 2022-08-05 #175 @So-Cool
Thanks so much for your comments, @kvarada.

- The paper is here: [`paper/paper.md`](https://github.com/fat-forensics/Surrogates-Tutorial/blob/master/paper/paper.md).
- Good catch regarding the licences. I have now fixed it and placed two badges in the README file and all of the Jupyter Notebooks.
- I have also supplemented the README file with additional details.
- The materials are focused on a toy example to facilitate an intuitive explanation of the relevant concepts without polluting the story with domain- and data-specific issues; an exploration of a real-life data set is left to the tutees as an exercise. Since each data set is unique there is no real advantage of discussing any particular scenario in detail. Nonetheless, the hope is that any specific exploration based on this collection of educational resources would link back to this repository, therefore building up a directory of problem-specific tutorials on composing bespoke explainers.
- I've added a few pointers to other (educational) materials that show how generic (one-size-fits-all) explainers may produce subpar explanations at the end of the last Jupyter Notebook (`4-explanation-generation.ipynb`).
