### 2022-08-13 de10bdd
(PF) minor textual changes
