### 2022-08-14 2701550
Learning outcomes
