### 2022-08-15 49bb18d
Paper: articulate learning outcomes & update affiliations
