### 2022-08-15 #175 @So-Cool
To clarify the learning outcomes, these are now listed in the *Statement of Need* section.
