### 2022-08-15 #175 @arokem
I have now finally had a chance to read the paper thoroughly and to go through all of the materials. I have checked off all of the boxes in my check-list and I am happy to support publication of this paper.
