### 2022-08-25 #175 @labarba
@kvarada – you have one unchecked item in [your list](https://github.com/openjournals/jose-reviews/issues/175#issuecomment-1200674854)... did you have a concern here, or you missed it?

Are you satisfied with the author's responses to your comments, at this point? Many thanks!
