+++ paper/paper.bib
@@ -32,13 +32,16 @@
   year={2019}
 }
 
-@article{sokol2019fatf,
+@article{sokol2022fatf,
   title={{FAT Forensics}: {A} {P}ython Toolbox for Algorithmic Fairness,
          Accountability and Transparency},
   author={Sokol, Kacper and Santos-Rodriguez, Raul and Flach, Peter},
-  journal={{arXiv} preprint {arXiv}:1909.05167},
-  doi={10.48550/arXiv.1909.05167},
-  year={2019}
+  journal={Software Impacts},
+  volume={14},
+  pages={100406},
+  issn={2665-9638},
+  doi={10.1016/j.simpa.2022.100406},
+  year={2022}
 }
 
 @article{sokol2020fatf,
