### 2022-09-08 #175 @So-Cool
The paper has been updated to include up-to-date references and acknowledgements.

@labarba, what are the next steps?
