### 2022-10-19 #175 @ttimbers
@So-Cool - we have two things left for you to do (or perhaps you have done them and you just need to tell me about them):

1. Make a tagged release on the target repository, and report the version number here.
2. Make an archive of the repository on Zenodo or an equivalent service, and report the archive DOI here.
