### 2022-10-20 #175 @So-Cool
Hi @ttimbers, here are the details you requested:

1. GitHub release tag is `2022-jose`;
2. Zenodo DOI is `10.5281/zenodo.7229153`.
