### 2022-10-20 #175 @ttimbers
Thanks @So-Cool. I think the version needs to be done using semantic versioning however...  So something like v1.x.x - do you think you can update the version to something like that?
