### 2022-10-21 #175 @So-Cool
No worries, @ttimbers. I've updated the releases to use semantic versioning. The latest release is tagged with `1.0`.
