### 2022-10-21 #175 @ttimbers
@openjournals/jose-eics - I think this paper should now be ready for full acceptance. Please let me know if I missed a step with whedon.
