### 2022-12-06 #175 @labarba
Hi @So-Cool — Apologies for the delay in executing the last step for publication. That is on me. I have read your proof one last time, and will proceed now.
