### 2022-12-06 #175 @labarba
Congratulations, @So-Cool, your JOSE paper is now published! 🚀

Huge thanks to the Editor: @ttimbers and Reviewers: @kvarada, @arokem — your contributions are invaluable 🙏
